Sci_Fi_Shooter_V1
As off this version I've got the base animations working for the current character (Able to move, and jump, plus both the lower and top halves of the body animate independently so that it can still shoot and jump). I Also have the first gun model ready. My next goal is to try and get it to work with the mixamo model.