Sci_Fi_Shooter_Weapons_and_Flight_update
I've manage to implement shooting for both the pistol and the assault rifle, and even though it's not working yet, and ricochet blaster is a WIP. I've also added limited ammo to the weapons. I also added in the ability to fly, along with a limit on how long you can flying for, which will recharge when your not flying (Note: You activate and disable flying by double tapping space/jump).