Animation, shooting and Damage update
I've made it so that the characters upper idle/iron sight poses change depending on the currently equipped weapon (And created a new animation for iron sight aiming with a pistol), and have managed to make it so that the 3 guns create a sound when fired. The pistol and assault rifle can also now do damage when there being a test target in the game to test damage and killing enemies. The only other issue I have right now besides getting the ricochet blaster to shoot properly is adding/finding a muzzle flash animation for the weapons.