Shooting mechanic update
Managed to not only get the Ricco gun to properly shoot ricocheting projectiles (Which also deal damage), but I also tweaked the aiming slightly. Now when you aim and fire, a hit scan is fired from the center of the camera, and the weapon hit scans will now hit where the camera hit scan landed. Allowing for more accurate aiming.