Running Update
I've implemented a mechanic that allows the user to run, allowing them to navigate the level more easily. This works both in the air and when flying. But I will need to rework the animation BP at a later date. I'm currently working on adding in the speed character (Exo red).