Exo red movement/weapon update
I reworked the movement speeds for exo red to make him generally faster than ely, and I also coded both the smg and the sniper (Both with sounds and the ladder has a 3 second cooldown before it can be fired again). I also decided to make the original pistol (Ely weapon) sound effect the smg (Exo red) sound effect and added a new sound for the pistol. The reason is because I feel a lighter sound "Pew" sound would work better for a faster but weaker weapon, and the pistol having a slightly louder/heavier sound is more fitting for it.