Weapon and character update
I finished the rail gun for the vanguard so now it shoots 1 of 4 different colored projectiles based on the amount of charge (Each charge also consumes 10-40 ammo, and the charge won't increase if you don't have enough ammo). I also added a cooldown for the rocket launcher. I also decided to rework the ammo a little. Now all 5 ammo types are stored in the third person BP instead of just bullets and energy cells. I did this because I'll think it'll help with setting up the UI. But I made it so the amount of each ammo is set on begin play based on the character (Exo-red has less bullets due to his strongest weapon,l the sniper using bullets), along with setting each characters Max and current HP (Another 2 floats I set up in the third person blueprint class). I also made a damage tester to test the characters taking damage, along with making them set to ragdoll physics on death. Lastly, I made it so when ironsight aiming as Exo reds with the sniper equiped, you can click the middle mouse butter to toggle sniper/zoomed in aim, and it all resets to normal if you release right click/iron sight aim. Along with making it so you can't switch weapons when ironsight aiming to avoid any issues like being able to sniper zoom with a shotgun.