Variable/flight restore update
I made the variables for flight time larger (Basically make it the original multiplies by 60 so that it can still last the original time while drain by 1 every frame), in order to help make the meter for flight time in the UI look alot smoother while it refulls and drains (Planning ahead for the UI). This also involved setting the variables on event begin play for each child class.