Weapon and UI update
I reworked the aiming for the projectile based weapons to make them aim more towards the center of the screen. I also added in the UI which displays a bar for both health and the current flight time remaining. And It also not only displays the current ammo for the equipped weapon, but it also shows the max bullets and energy cells for the weapons that use them (I made a mx bullets and max energy cells variable which is also changed depending on the character you start with).