UI updates
I made it so the aim scope only appears when aiming, and will be switched with a sniper scope when sniper aiming as exo red (Note: I do intend to create a more robust UI for when your sniper aiming, but am experiencing technical difficulties right now). Secondly, I added the ability to pause the game, currently only the resume button works, but I'll add the other functions as time goes on.