Weapon and UI work in progress
I'm in the process of developing the character selection menu (Have also set up game instance), and currently need to try to figure out how to change the character you spawn as when you start a level. I'm also reworking the code for the aiming to make it more reliable (Sometimes the damage line trace fires off to the corner).