UI and level start up work in progress
I streamlined the widget switching by setting the variable in the third person character parent class and using that in the child classes, but I still need to work on getting the level to start properly. Mainly with setting the right default widget, and getting the line trace from the camera to work properly.