Small options update
I added in a volume slider, a mute button/checkbox and a percentage display button/checkbox (This will display the percentage of health and flight time remaining), along with variables for these respective options in the game instance. As of right now I have it so the audio variables affect the assault rifle Ely uses, will add it to the other weapons in the next update.