Menu, UI and weapon update
I finished up adding the mute audio and volume controls to all the sound effects. Meaning now you can adjust the sound and audio in the options. I also added a show percentage button, which will display the current percentage (out of 100) for both health and flight time based on an option in the settings. Lastly, I reworked the aiming for the projectile based weapons (Ely's riccochet gun and all 3 of the vanguards weapons), to make them aim towards the center of the screen/crosshairs more accuratly)