UI and weapon projectile update
First to correct a mistake from the last push, I not only added in more commenting, but also added hit particles for the rail guns projectiles. For this update, I reworked the menus to be more centered on the screen. make it so the bullet effect for hit scan weapons disapears on hit, and reduced the sniper cooldown from 3 seconds to 1.4 seconds (Sets the timer to 100 which goes down by 1 every frame, which rounds to 1.4 seconds if the framerate is 60 FPS).